+++ mustang_codegen/lib/builder.dart
@@ -1,6 +1,7 @@
 import 'package:build/build.dart';
 import 'package:mustang_codegen/src/app_serializer_builder.dart';
 import 'package:mustang_codegen/src/aspect_generator/app_aspect_generator.dart';
+import 'package:mustang_codegen/src/batch_generator/app_batch_generator.dart';
 import 'package:mustang_codegen/src/model_generator/app_model_generator.dart';
 import 'package:mustang_codegen/src/screen_generator/screen_generator.dart';
 import 'package:mustang_codegen/src/service_generator/screen_service_generator.dart';
@@ -10,6 +11,11 @@ import 'package:source_gen/source_gen.dart';
 // Unused for now
 // /// Generates mustang config
 // Builder mustangConfigBuilder(BuilderOptions options) => MustangConfigBuilder();
+
+Builder appBatchLibraryBuilder(BuilderOptions options) => LibraryBuilder(
+      AppBatchGenerator(),
+      generatedExtension: '.batch.dart',
+    );
 
 /// Generates code for @Around, @Before, @After annotations
 Builder appAspectLibraryBuilder(BuilderOptions options) => LibraryBuilder(