+++ mustang_codegen/lib/src/batch_generator/batch_model.dart
@@ -1,0 +1,12 @@
+import 'package:analyzer/dart/element/element.dart';
+
+class BatchModel {
+  const BatchModel({
+    required this.serviceClass,
+    required this.methodWithBatchRun,
+  });
+
+  final ClassElement serviceClass;
+
+  final MethodElement methodWithBatchRun;
+}