+++ mustang_core/lib/mustang_core.dart
@@ -6,6 +6,8 @@ export 'src/annotations/app_event_global.dart';
 export 'src/annotations/app_model.dart';
 export 'src/annotations/around.dart';
 export 'src/annotations/aspect.dart';
+export 'src/annotations/batch.dart';
+export 'src/annotations/batch_run.dart';
 export 'src/annotations/before.dart';
 export 'src/annotations/init_field.dart';
 export 'src/annotations/invoke.dart';